Updated Item Textures, Used the Structure template for Amenonuhoko, Fixed some Items, Updated Rod of Remorse, Updated Hakurei Gohei, Fixed Onbashira replacing blocks when placed, Fixed Ghost Dipper not working properly.

+++ src/main/java/arekkuusu/grimoireofalice/common/item/ItemSacredToyosatomimi.java
@@ -89,7 +89,7 @@ public class ItemSacredToyosatomimi extends ItemSwordOwner {
 						player.addChatComponentMessage(
 								new TextComponentString(TextFormatting.GOLD + "- - - - - - - - - - - - - - - - - - - - - - - - -"));
 						list.forEach(mob -> player.addChatComponentMessage(new TextComponentString(
-								TextFormatting.GOLD + "- " + mob.getName() + " : {" + mob.posX + ", " + mob.posY + ", " + mob.posZ + "}")));
+								TextFormatting.GOLD + "- " + mob.getName() + " : {" + (int)mob.posX + ", " + (int)mob.posY + ", " + (int)mob.posZ + "}")));
 						player.addChatComponentMessage(
 								new TextComponentString(TextFormatting.GOLD + "- - - - - - - - - - - - - - - - - - - - - - - - -"));
 					}
